requirements.txt
# pypotrace
matplotlib~=3.9.4
numpy~=1.26.4
scipy~=1.13.1
stringzilla~=3.12.5
jax~=0.4.30
jaxlib~=0.4.30
opt-einsum~=3.4.0
absl-py~=2.2.2
flatbuffers~=25.2.10
protobuf~=4.25.6
ml-dtypes~=0.5.1
colorama~=0.4.6
requests~=2.32.3
pillow~=11.2.1
pip~=25.0.1
attrs~=25.3.0
cffi~=1.17.1
setuptools~=58.1.0
onnx~=1.17.0
packaging~=24.2
PyYAML~=6.0.2
zipp~=3.21.0
llvmlite~=0.43.0
numba~=0.60.0
importlib-metadata~=8.6.1
Cython~=3.0.12
pytz~=2025.2
charset-normalizer~=3.4.1
pooch~=1.8.2
tqdm~=4.67.1
platformdirs~=4.3.7
PyMatting~=1.1.13
jsonschema~=4.23.0
threadpoolctl~=3.6.0
joblib~=1.4.2
pyserial~=3.5
certifi~=2025.1.31
imageio~=2.37.0
tifffile~=2024.8.30
scikit-image~=0.24.0
lazy-loader~=0.4
networkx~=3.2.1
scikit-learn~=1.6.1
pandas~=2.2.3
urllib3~=2.4.0
wcwidth~=0.2.13
simsimd~=6.2.1
albucore~=0.0.23
six~=1.17.0
python-dateutil~=2.9.0.post0
pydantic~=2.11.3
annotated-types~=0.7.0
idna~=3.10
contourpy~=1.3.0
fonttools~=4.57.0
mediapipe~=0.10.21
sounddevice~=0.5.1
sentencepiece~=0.2.0
pycparser~=2.22
pyparsing~=3.2.3
referencing~=0.36.2
jsonschema-specifications~=2024.10.1
importlib-resources~=6.5.2
cycler~=0.12.1
kiwisolver~=1.4.7
albumentations~=2.0.5
prettytable~=3.16.0
eval-type-backport~=0.2.2
openpyxl~=3.1.5
et-xmlfile~=2.0.0